requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
pydantic-settings 
mStock-TradingApi-A